整理测试, janim 适配框架
