+++ tests/janim_anime/tan.py
